==> notebooks/1-exploracion_inicial.ipynb <==
# Fase 1: Unión de Conjuntos de Datos

### 1. Lectura y Exploración Inicial

import pandas as pd
import numpy as np
import json
import os 
import re

import sys 
sys.path.append("../")
from src import soporte_limpieza

os.chdir("../datos")
datos = os.listdir()[:9]
datos

Creamos una lista con los df

dic_df ={}
lista_anios = []

for i in datos:
    df = pd.read_csv(f"../datos/{i}", sep=";", parse_dates=["DATA LANÇAMENTO"])
    # print(i)
    año = re.findall("\d{4}", i)[0]
    dic_df[int(año)] = df
    lista_anios.append(int(año))


# Observamos el formato de los dataframes
for elem in lista_anios:
    print(dic_df[elem].shape)

Se observa que el volumen de datos es mucho menor en los 3 primeros años, pero todos tienen el mismo número de columnas. Vamos a cambiar el nombre de las columnas de todos los data frames para facilitar la comprensión.

cambio_nombre_columnas = {"CÓDIGO ÓRGÃO SUPERIOR":"id organizacion superior",
                          "NOME ÓRGÃO SUPERIOR":"organizacion superior",
                          "CÓDIGO ÓRGÃO":"id organizacion",
                          "NOME ÓRGÃO":"organizacion",
                          "CÓDIGO UNIDADE GESTORA":"id unidad gestora",
                          "NOME UNIDADE GESTORA":"unidad gestora",
                          "CATEGORIA ECONÔMICA":"categoria economica",
                          "ORIGEM RECEITA":"origen ingreso",
                          "ESPÉCIE RECEITA":"tipo ingreso",
                          "DETALHAMENTO":"detalles",
                          "VALOR PREVISTO ATUALIZADO":"valor previsto", #Este valor era el previsto que se iba a recaudar
                          "VALOR LANÇADO":"valor registrado", #Este valor es el que han apuntado oficialmente como recaudado 
                          "VALOR REALIZADO":"valor recaudado", #Este valor es el que realmente han recaudado
                          "PERCENTUAL REALIZADO":"porcentaje recaudado", #Porcetanje del valor recaudado respecto al previsto
                          "DATA LANÇAMENTO":"fecha recaudacion",
                          "ANO EXERCÍCIO":"anio recaudacion"
                          }

for elem in lista_anios:
    dic_df[elem].rename(columns=cambio_nombre_columnas, inplace=True)

dic_df[2019].head(2)

No tenía muy claro si el año de la columna fecha recaudacion coincide con el de la columna anio recaudacion, en gan parte de los casos era asi y he querido ver donde no coincidian descubriendo que hay una gran cantidad de valores en Nan, pensé en asignar al año de recaudacion el año de la columna fecha recaudacion pero hay filas donde ambos son Nan, por lo que no es una buena estrategia. Como hemos vito que solo hay un valor único que es el año que coincide co el nombre del csv, podemos sustituir todos los Nan del año de recaudación por el año del csv. Los Nan de la fecha de recaudación los dejamos sin tratar y si fuera necesario en el futuro ya veremos que hacemos con ellos.

dic_df[2013].shape

dic_df[2013].dtypes

filtro = pd.to_datetime(dic_df[2013]["fecha recaudacion"]).dt.year != dic_df[2013]["anio recaudacion"]
df_filtro = dic_df[2013][filtro]
df_filtro.head(1)

dic_df[2013][dic_df[2013]["anio recaudacion"].isnull() & dic_df[2013]["fecha recaudacion"].isnull()].head(1)

for elem in lista_anios:
    print(dic_df[elem]["anio recaudacion"].unique())

for anio in lista_anios:
    dic_df[anio]["anio recaudacion"]=dic_df[anio]["anio recaudacion"].fillna(anio)

for elem in lista_anios:
  print(dic_df[elem]["anio recaudacion"].unique())

---------------------------

Se observó antes que los tipos de columnas numércicas como son, valor previsto, registrado y recaudado son de tipo object cuano deberían ser un float,cambiaremos los valores a float.

for anio in lista_anios:
    print(f"\n{anio}")
    print(dic_df[anio][["valor previsto", "valor registrado", "valor recaudado", "porcentaje recaudado"]].dtypes)

# Para convertirlo a float primero tenemos que sustituir la coma por un punto
columnas = ["valor previsto", "valor registrado", "valor recaudado", "porcentaje recaudado"]

for anio in lista_anios:
    for columna in columnas: 
        dic_df[anio][columna]=dic_df[anio][columna].str.replace(",", ".")

dic_df[2021][["valor previsto", "valor registrado", "valor recaudado", "porcentaje recaudado"]].dtypes

# Convertimos las columnas a tipo float
columnas = ["valor previsto", "valor registrado", "valor recaudado", "porcentaje recaudado"]
for anio in lista_anios:
    for column in columnas:
        dic_df[anio][column]=dic_df[anio][column].astype(float)

for anio in lista_anios:
    print(dic_df[anio][["valor previsto", "valor registrado", "valor recaudado", "porcentaje recaudado"]].dtypes)

Todas las columnas numéricas de las tablas han sido satisfactoriamente convertidas a float.

Explorando estos datos para hacer la conversión me he fijado en que algunos valores estaban a 0 de forma recurrente, indaguemos un poco más en esto.

columnas = ["valor previsto", "valor registrado", "valor recaudado"]
for anio in lista_anios:
    print(f"\n{anio}")
    for columna in columnas:
        print(columna)
        print(f"  {(dic_df[anio][columna].value_counts()/dic_df[anio].shape[0]*100)[0]}")

Estas columnas son de vital importancia ya que reflejan cuando dinero se preve obtener, cuanto se ha regitrado y cuanto se ha recaudado, por lo que ver tantos datos a 0 indica un grave problema burocrático a la hora de llevar las cuentas. Estos valores a 0 se gestionarán en el próximo notebook.

________________

Procedemos a unir las tablas y asi trabajar sobre una única tabla. Los nombres de las columnas y su formato es homogéneo por lo que podemos proceder a la concatenación, método seleccionado ya que cada tabla tiene las mismas columnas y la única opción con sentido es unir las tablas una debajo de otra.

for anio in lista_anios:
    print(dic_df[anio].dtypes)

lista_df=[]
for anio in lista_anios:
    lista_df.append(dic_df[anio])

suma = 0
for anio in lista_anios:
    suma += dic_df[anio].shape[0]
suma

df_concat = pd.concat(lista_df, axis=0, ignore_index = True)
df_concat.shape


Comprobamos satisfactoriamente que la concatenación se ha realizado comprobando que el número de filas totales es la suma de las filas de cada tabla. Eliminamos filas duplicadas, pasamos la fecha a formato date, elminamoscolumnas innecesarias y guardamos a nuestro pequeño monstruo en un nuevo csv.

df_concat.drop_duplicates(inplace=True)

df_concat.shape

df_concat.dtypes

df_concat["fecha recaudacion"]= pd.to_datetime(df_concat["fecha recaudacion"])

df_concat.dtypes

df_concat.drop("detalles", axis = 1, inplace = True) #Columnna inecesaria por lo que la quitamos

df_concat.columns

df_concat.to_csv("../datos/tablas_concatenadas.csv")

df_concat.isnull().sum()
