==> notebooks/2-limpieza.ipynb <==
# Fase 2: Limpieza de Datos

import pandas as pd
import numpy as np
import json

import sys
import os

# Agregar el directorio 'src' al path
sys.path.append(os.path.abspath('../src'))

# Importar la función suma del módulo auxiliares
import soporte_limpieza as sl

df = pd.read_csv("../datos/tablas_concatenadas.csv", parse_dates=["fecha recaudacion"])
df.head(2)

df.isnull().sum()

### 1- Gestión de los nulos de id organizacion superior y organizacion superior.

Vamos a ver si podemos crear un diccionario para los id y los nombres de la organización superior.

len(df["id organizacion superior"].unique())

len(df["organizacion superior"].unique())

A priori parece que hay el mismo numero de ids que de nombres, buena señal.

lista_id_os = list(df["id organizacion superior"].unique())
lista_id_os

for id in lista_id_os:
    print(f"\n{id}")
    filtro1 = df["id organizacion superior"] == float(id)
    filtro2 = df["organizacion superior"].notnull()
    print(len(df[filtro1 & filtro2]["organizacion superior"].unique()))

Como vemos que para cada id hay un unico valor creamos un diccionario:

dic_id_os = {}
for id in lista_id_os:
    try:
        filtro1 = df["id organizacion superior"] == float(id)
        filtro2 = df["organizacion superior"].notnull()
        dic_id_os[float(id)]=df[filtro1 & filtro2]["organizacion superior"].unique()[0]
    except:
        pass

dic_id_os

Ya con nuestro diccionario vamos a rellenar los nombres e id de las organizaciones superiores.

Hemos comprobado que todos los campos de organización superior concuerdan con su id a excepción de cuando son NaN, corregimos los NaN. además quedarán sin corregir aquellas filas donde id y organización superioir son ambos NaN.

# filas que van a quedar sin corregir.
filtro1 = df["id organizacion superior"].isnull()
filtro2 = df["organizacion superior"].isnull()
df[filtro1 & filtro2].shape

df["organizacion superior"].fillna(df["id organizacion superior"].map(dic_id_os), inplace=True)

os_null = df[df["organizacion superior"].isnull()].shape[0]
os_sin_corregir = os_null/df.shape[0]*100
print(f"Han quedado a null {os_null}, es decir, un {np.round(os_sin_corregir,2)}% del total de las organizaciones superiores.")

df[df["id organizacion superior"].isnull()].shape[0]

Como podemos ver ids hay 30359 a null los cuales podemos corregir al rededor de unos 20000 a partir del nombre de la organizacion superior.

#Invertimos el diccionario
dic_organizaciones_id = {valor: clave for clave, valor in dic_id_os.items()}
dic_organizaciones_id


df["id organizacion superior"].fillna(df["organizacion superior"].map(dic_organizaciones_id), inplace=True)

df[df["id organizacion superior"].isnull()].shape[0]

Comprobamos cuantos null hay ahora y efectivamente se han reducido hasta 10545, los mismo que para los nombres de las organizaciones superiores ya que son todas las filas que tienen ambos campos a null. Aunque al representar un 1% y al no haber forma de arreglarlo puede ser despreciable.

df.isnull().sum()

___

# pd.set_option("display.max_rows", None)
resultado = df.groupby(['id organizacion'])['id organizacion superior'].nunique().reset_index()
resultado[resultado["id organizacion superior"]!=1]

Por lo tanto para cada par de id orgnización solo va a estar asociado a un único id organizacion superior. Creamos un diccionario con clave el id de organización y como valor tuplas de los valores de id organizacion superior y organizacion superior.

# Eliminamos aquellas filas donde haya nulos
df_validos = df.dropna(subset=['id organizacion superior', 'organizacion superior', 'id organizacion'])

diccionario_orga_a_sup = {}
for index, row in df_validos.iterrows():
    diccionario_orga_a_sup[row['id organizacion']] = (row['id organizacion superior'], row['organizacion superior'])

len(diccionario_orga_a_sup)

with open('../datos/diccionario_orga_a_sup.json', 'w') as file:
    json.dump(diccionario_orga_a_sup, file)

He intentado meter esta funcion en el archivo .py pero llevaba más de dos minutos ejecutandose y no terminaba, en cambio si la dejo aquí tarda 30 secs.

def rellenar_orga_sup(row):
    # Accedemos a los nulos que etán donde ambos campos son nulos.
    if pd.isna(row['id organizacion superior']) and pd.isna(row['organizacion superior']):
        id_org = row['id organizacion']
        if id_org in diccionario_orga_a_sup:
            row['id organizacion superior'], row['organizacion superior'] = diccionario_orga_a_sup[id_org]
    return row

df = df.apply(rellenar_orga_sup, axis=1)

df.isnull().sum()/df.shape[0]*100

Hemos conseguido reducir los nulos a un 0.03% para id organizacion superioir y organizacion superior.

### 2- Gestión de los nulos de id organizacion y organizacion.

Vamos a seguir la misma estrategia que en el paso anterior.

len(df["id organizacion"].unique())

len(df["organizacion"].unique())

Hay más ids que organizaciones, lo cual impica que una misma organización puede tener 2 id, veámoslo.

lista_organizaciones = list(df["organizacion"].unique())
len(lista_organizaciones)

for orga in lista_organizaciones:
    filtro1 = df["id organizacion"].notnull()
    filtro2 = df["organizacion"] == orga
    if len(df[filtro1 & filtro2]["id organizacion"].unique()) > 1:
        print(f"\n{orga}")
        print(df[filtro1 & filtro2]["id organizacion"].unique())

Efectivamente vemos que hay tres organizaciones con dos id. Vamos a ver que está pasando.

- Para la Agência Nacional de Transportes Aquaviários vemos que el id 39251 solo seusa en 5 filas y todas son de 2013 por lo que puede ser que sea un id viejo y haya transitado hacia el nuevo que es 68201, por lo que vamos a cambiar el id viejo por el nuevo para homogeneizar los datos.

filtro1_1 = df["id organizacion"] == 68201
df[filtro1_1].shape
df[filtro1_1].head(1)

filtro1_2 = df["id organizacion"] == 39251
df[filtro1_2].shape
df[filtro1_2].head(1)

df["id organizacion"] = df["id organizacion"].apply(sl.cambio_id_39251)

filtro1 = df["id organizacion"] == 39251
df[filtro1].shape

- Para la Fundo Nacional Antidrogas pasa algo similar, sin embargo, el id 20117 aparece 20 veces y de forma aleatoria pues aparece en 2013, 2014, 2015, 2016 y hasta 2017, pero no a partir de 2017, lo que puede indicar que se fue transicionando de id durante varios años usandose dos a la vez hasta que en 2017 se dejó de usar. Además también se observa que para ambos ids a pesar de tener la misma unidad gestora también tienen ids de unidad gestora distintos.

IMPORTANTE:

He vuelto sobre mis pasos y me he dado cuenta de que la organización superior para los distintos id es distinta, por lo que si los unificara todos es como si estuviera perdiendo información ya que aunque la organización sea la misma si tiene el id de organización 30912 quiere decir que su organización superior es Ministério da Justiça e Segurança Pública y si tiene el id 20117 su organización superioir es Presidência da República. Me he dado cuenta de que pasa lo mismo también para los dos siguientes. En mi opinión unificar ambos id y dejar el que más aparece implicaría en cierta forma perder información, ya que si yo tengo un valor NaN para orga superioir e id de orga superior pero con id de organización 20117 se puede inferir que la organización superior es Presidência da República. Pero si unificasemos los id y pusiesemos para la organización Fundo Nacional Antidroga el id 30912 si tenemos un valor en NaN para id de orga superior y orga superior no podemos inferir si era Presidência da República o Ministério da Justiça e Segurança Pública. Pero como hemos visto que solo tenemos a NaN un 1% de las filas esto lo gestionaré si tuviera tiempo más adelante.

filtro2_1 = df["id organizacion"] == 30912
print(df[filtro2_1].shape)
print(df[filtro2_1]["id unidad gestora"].unique())
print(df[filtro2_1]["unidad gestora"].unique())
df[filtro2_1].head(1)

filtro2_2 = df["id organizacion"] == 20117
print(df[filtro2_2].shape)
print(df[filtro2_2]["id unidad gestora"].unique())
print(df[filtro2_2]["unidad gestora"].unique())
df[filtro2_2].head(1)

df["id organizacion"] = df["id organizacion"].apply(sl.cambio_id_20117)

filtro2 = df["id organizacion"] == 20117
df[filtro2].shape

- Continuamos con los id de Fundo Nacional do Idoso. De nuevo ocurre lo mismo, el id 64902 aparece solo 99 veces y hasta 2017, el id 30914 aparece 223 y a partir de 2017 lo que indica un camio de id, siendo el actual 30914. De nuevo los id de unidad gestora son distintos aunque la unidad gestora sea la misma.


filtro3_1 = df["id organizacion"] == 64902
print(df[filtro3_1].shape)
df[filtro3_1].head(1)

filtro3_2 = df["id organizacion"] == 30914
print(df[filtro3_2].shape)
df[filtro3_2].head(1)

df["id organizacion"] = df["id organizacion"].apply(sl.cambio_id_64902)

filtro3 = df["id organizacion"] == 64902
df[filtro2].shape

- Terminamos con Fundo Nacional para a Criança e o Adolescente, en este caso ambos id se usn en 2021, el 64901 se empieza a usar en 2013 mientras que es 30913 se usa a apartir de 2017, por lo que vamos a dejar 30913.

filtro4_1 = df["id organizacion"] == 64901
print(df[filtro4_1].shape)
filtro4_2 = df["id organizacion"] == 30913
print(df[filtro4_2].shape)

df[filtro4_2].head(1)

df[filtro4_1].head(1)

df["id organizacion"] = df["id organizacion"].apply(sl.cambio_id_64901)

filtro4 = df["id organizacion"] == 64901
df[filtro4].shape

#Comprobamos que ya solo hay un id por organización
flag = 0
for orga in lista_organizaciones:
    filtro1 = df["id organizacion"].notnull()
    filtro2 = df["organizacion"] == orga
    if len(df[filtro1 & filtro2]["id organizacion"].unique()) > 1:
        flag=1
        print(f"\n{orga}")
        print(df[filtro1 & filtro2]["id organizacion"].unique())
print(flag)

Ya podemos hacer un diccionario para rellenar valores.

lista_id_organizaciones = list(df["id organizacion"].unique())
print(len(lista_id_organizaciones))

dic_id_organizacion = {}
for id in lista_id_organizaciones:
    try:
        filtro1 = df["id organizacion"] == float(id)
        filtro2 = df["organizacion"].notnull()
        dic_id_organizacion[float(id)]=df[filtro1 & filtro2]["organizacion"].unique()[0]
    except:
        pass

df["id organizacion"].isnull().sum()

# filas que van a quedar sin corregir.
filtro1 = df["id organizacion"].isnull()
filtro2 = df["organizacion"].isnull()
df[filtro1 & filtro2].shape

df["organizacion"].fillna(df["id organizacion"].map(dic_id_organizacion), inplace=True)

orga_null = df[df["organizacion"].isnull()].shape[0]
orga_sin_corregir = orga_null/df.shape[0]*100
print(f"Han quedado a null {orga_null}, es decir, un {np.round(orga_sin_corregir,2)}% del total de las organizaciones.")

df[df["id organizacion"].isnull()].shape[0]

Como podemos ver ids hay 25117 a null los cuales podemos corregir al rededor de unos 24200 a partir del nombre de la organizacion.

#Invertimos el diccionario
dic_orga_id = {valor: clave for clave, valor in dic_id_organizacion.items()}

df["id organizacion"].fillna(df["organizacion"].map(dic_orga_id), inplace=True)

df[df["id organizacion"].isnull()].shape[0]

Comprobamos cuantos null hay ahora y efectivamente se han reducido hasta 905, los mismo que para los nombres de las organizaciones ya que son todas las filas que tienen ambos campos a null. Aunque al representar un 0.1% y al no haber forma de arreglarlo puede ser despreciable.

### 3- Gestión de los nulos de id unidad gestora y unidad gestora.

Vamos a seguir la misma estrategia que en el paso anterior.

len(df["id unidad gestora"].unique())

len(df["unidad gestora"].unique())

Como esperábamos, hay más ids que unidades gestoras.

lista_gestoras = list(df["unidad gestora"].unique())
len(lista_gestoras)

for gestora in lista_gestoras:
    filtro1 = df["id unidad gestora"].notnull()
    filtro2 = df["unidad gestora"] == gestora
    if len(df[filtro1 & filtro2]["id unidad gestora"].unique()) > 1:
        print(f"\n{gestora}")
        print(df[filtro1 & filtro2]["id unidad gestora"].unique())

- Para SETORIAL PROG.ORCAMENTARIA E FINANCEIRA-SFB, el id 130214 solo aparece 7 veces en 2019, lo reemplazamos.

filtro1_1 = df["id unidad gestora"] == 440088
print(df[filtro1_1].shape)
df[filtro1_1].head(1)

filtro1_2 = df["id unidad gestora"] == 130214
print(df[filtro1_2].shape)
df[filtro1_2].head(1)


df["id unidad gestora"] = df["id unidad gestora"].apply(sl.cambio_id_130214)

filtro1 = df["id unidad gestora"] == 130214
df[filtro1].shape

A partir de aqui he implementado la estrategia de que el id que aparezca en más filas es el que prevalece, me gustaría pararme a analizarlo en mayor profundidad pero no me da tiempo.

- COORDENACAO-GERAL DE ORCAMENTO E FINANCAS

filtro2_1 = df["id unidad gestora"] == 240102
print(df[filtro2_1].shape)
df[filtro2_1].head(1)

filtro2_2 = df["id unidad gestora"] == 410002
print(df[filtro2_2].shape)
df[filtro2_2].head(1)

df["id unidad gestora"] = df["id unidad gestora"].apply(sl.cambio_id_410002)

filtro2 = df["id unidad gestora"] == 410002
df[filtro2].shape


- DIRETORIA DE FINANCAS-SISTEMA PARA O PAIS/MM


filtro3_1 = df["id unidad gestora"] == 673001
print(df[filtro3_1].shape)
df[filtro3_1].head(1)

filtro3_2 = df["id unidad gestora"] == 873001
print(df[filtro2_2].shape)
df[filtro3_2].head(1)

df["id unidad gestora"] = df["id unidad gestora"].apply(sl.cambio_id_873001)

filtro3 = df["id unidad gestora"] == 873001
df[filtro3].shape

- AGENCIA NACIONAL DE TRANSPORTES AQUAVIARIOS


filtro4_1 = df["id unidad gestora"] == 682010
print(df[filtro4_1].shape)
df[filtro4_1].head(1)

filtro4_2 = df["id unidad gestora"] == 393002
print(df[filtro4_2].shape)
df[filtro4_2].head(1)

df["id unidad gestora"] = df["id unidad gestora"].apply(sl.cambio_id_393002)

filtro4 = df["id unidad gestora"] == 393002
df[filtro4].shape

- FUNDO NACIONAL ANTIDROGAS


filtro5_1 = df["id unidad gestora"] == 200246
print(df[filtro5_1].shape)
df[filtro5_1].head(1)

filtro5_2 = df["id unidad gestora"] == 110246
print(df[filtro5_2].shape)
df[filtro5_2].head(1)

df["id unidad gestora"] = df["id unidad gestora"].apply(sl.cambio_id_110246)

filtro5 = df["id unidad gestora"] == 110246
df[filtro5].shape

- FUNDO NACIONAL DO IDOSO


Me he dado cuenta de que esto lo podía haber automatizado antes y de una forma aún más eficiente, pero como ya me quedan tres no voy a cambiarlo todo.

num1=207001
num2=307002

filtro1 = df["id unidad gestora"] == num1
valor1 = df[filtro1].shape[0]
print(f"Para el {num1} tenemos un shape de {df[filtro1].shape}")

filtro2 = df["id unidad gestora"] == num2
valor2 = df[filtro2].shape[0]
print(f"Para el {num2} tenemos un shape de {df[filtro2].shape}")

if valor1 > valor2:
    print(f" Cambiamos {num2} --> {num1}")
    df["id unidad gestora"] = df["id unidad gestora"].apply(lambda id: num1 if id == num2 else id)
    filtro = df["id unidad gestora"] == num2
    print(df[filtro].shape)

else:
    print(f" Cambiamos {num1} --> {num2}")
    df["id unidad gestora"] = df["id unidad gestora"].apply(lambda id: num2 if id == num1 else id)
    filtro = df["id unidad gestora"] == num1
    print(df[filtro].shape)



- FUNDO NACIONAL PARA A CRIANCA E O ADOLESCENTE

num1=110244
num2=307001

filtro1 = df["id unidad gestora"] == num1
valor1 = df[filtro1].shape[0]
print(f"Para el {num1} tenemos un shape de {df[filtro1].shape}")

filtro2 = df["id unidad gestora"] == num2
valor2 = df[filtro2].shape[0]
print(f"Para el {num2} tenemos un shape de {df[filtro2].shape}")

if valor1 > valor2:
    print(f" Cambiamos {num2} --> {num1}")
    df["id unidad gestora"] = df["id unidad gestora"].apply(lambda id: num1 if id == num2 else id)
    filtro = df["id unidad gestora"] == num2
    print(df[filtro].shape)

else:
    print(f" Cambiamos {num1} --> {num2}")
    df["id unidad gestora"] = df["id unidad gestora"].apply(lambda id: num2 if id == num1 else id)
    filtro = df["id unidad gestora"] == num1
    print(df[filtro].shape)

- COORDENACAO-GERAL DE RECURSOS LOGISTICOS


num1=410003
num2=240101

filtro1 = df["id unidad gestora"] == num1
valor1 = df[filtro1].shape[0]
print(f"Para el {num1} tenemos un shape de {df[filtro1].shape}")

filtro2 = df["id unidad gestora"] == num2
valor2 = df[filtro2].shape[0]
print(f"Para el {num2} tenemos un shape de {df[filtro2].shape}")

if valor1 > valor2:
    print(f" Cambiamos {num2} --> {num1}")
    df["id unidad gestora"] = df["id unidad gestora"].apply(lambda id: num1 if id == num2 else id)
    filtro = df["id unidad gestora"] == num2
    print(df[filtro].shape)

else:
    print(f" Cambiamos {num1} --> {num2}")
    df["id unidad gestora"] = df["id unidad gestora"].apply(lambda id: num2 if id == num1 else id)
    filtro = df["id unidad gestora"] == num1
    print(df[filtro].shape)

Comprobamos que solo haya un id por unidad gestora:

flag=0
for gestora in lista_gestoras:
    filtro1 = df["id unidad gestora"].notnull()
    filtro2 = df["unidad gestora"] == gestora
    if len(df[filtro1 & filtro2]["id unidad gestora"].unique()) > 1:
        flag=1
        print(f"\n{gestora}")
        print(df[filtro1 & filtro2]["id unidad gestora"].unique())
print(flag)

Ya podemos hacer un diccionario para rellenar los valores:

lista_id_gestoras = list(df["id unidad gestora"].unique())
print(len(lista_id_gestoras))

dic_id_gestoras = {}
for id in lista_id_gestoras:
    try:
        filtro1 = df["id unidad gestora"] == float(id)
        filtro2 = df["unidad gestora"].notnull()
        dic_id_gestoras[float(id)]=df[filtro1 & filtro2]["unidad gestora"].unique()[0]
    except:
        pass

df["id unidad gestora"].isnull().sum()

# filas que van a quedar sin corregir.
filtro1 = df["id unidad gestora"].isnull()
filtro2 = df["unidad gestora"].isnull()
df[filtro1 & filtro2].shape

df["unidad gestora"].fillna(df["id unidad gestora"].map(dic_id_gestoras), inplace=True)

gestora_null = df[df["unidad gestora"].isnull()].shape[0]
gestora_sin_corregir = gestora_null/df.shape[0]*100
print(f"Han quedado a null {gestora_null}, es decir, un {np.round(gestora_sin_corregir,2)}% del total de las filas en unidades gestoras.")

df[df["id unidad gestora"].isnull()].shape[0]

Como podemos ver ids hay 33581 a null los cuales podemos corregir alrededor de unos 33000 a partir del nombre de la organizacion.

#Invertimos el diccionario
dic_gestora_id = {valor: clave for clave, valor in dic_id_gestoras.items()}

df["id unidad gestora"].fillna(df["unidad gestora"].map(dic_gestora_id), inplace=True)

df[df["id unidad gestora"].isnull()].shape[0]

Comprobamos cuantos null hay ahora y efectivamente se han reducido hasta 614, los mismo que para los nombres de las unidades gestoras ya que son todas las filas que tienen ambos campos a null. Aunque al representar un 0.06% y al no haber forma de arreglarlo puede ser despreciable.

### 4- Gestión de los nulos de categoría económica.

df.isnull().sum()/df.shape[0]*100

df["categoria economica"].value_counts()

df["categoria economica"].isna().sum()

df["categoria economica"].isna().sum()/df.shape[0]*100

Tenemos casi un 2% de nulos en esta columna, como no tenemos información con la que rellenarlos y sabemos que dentro de categorías exite una que es "Sem informação", en vez de dejar los nulos vamos a meterlos en esta categoría.

df["categoria economica"].fillna("Sem informação", inplace=True)

df["categoria economica"].value_counts()

df["categoria economica"].isna().sum()

### 5- Gestión de los nulos de origen ingreso.

df["origen ingreso"].unique()

df["origen ingreso"].isna().sum()/df.shape[0]*100

Tenemos un 3.8% de origenes de ingreso a nulo, como no tenemos información sidicinet como para saber de forma segura de donde vienen vamos a ponerlos como "Sem informação"

df["origen ingreso"].fillna("Sem informação", inplace=True)

df["origen ingreso"].isna().sum()

### 6- Gestión de los nulos de tipo ingreso.

df["tipo ingreso"].isna().sum()/df.shape[0]*100

De nuevo, en la columna "tipo ingreso" tenemos un 3.13% de los datos nulos, como hay muchas categorías y no sabems a cual pertenecen los nulos antes que darlos a NaN les metemos en "Sem informação".

df["tipo ingreso"].unique()

df["tipo ingreso"].fillna("Sem informação", inplace=True)

df["tipo ingreso"].isna().sum()

### 6- Gestión de los nulos de valores previsto, registrado, recaudado y porcentaje.


df.columns

df.isnull().sum()[["valor previsto", "valor registrado", "valor recaudado", "porcentaje recaudado"]]

(df.isnull().sum()/df.shape[0]*100)[["valor previsto", "valor registrado", "valor recaudado", "porcentaje recaudado"]]

filtro1 = df["valor previsto"].isnull()
filtro2 = df["valor registrado"].isnull()
filtro3 = df["valor recaudado"].isnull()
filtro4 = df["porcentaje recaudado"].isnull()

filtro5 = df["valor previsto"]==0
filtro6 = df["valor registrado"]==0
filtro7 = df["valor recaudado"]==0
filtro8 = df["porcentaje recaudado"]==0

filtros_comb = filtro1 & filtro2 & filtro3 & filtro4

print(df[filtro1].shape)
print(df[filtro2].shape)
print(df[filtro3].shape)
print(df[filtro4].shape)
print(f" Los 3 campos a null: {df[filtro1 & filtro2 & filtro3].shape}")
print(f" Los 4 campos a null: {df[filtros_comb].shape}")

- Para las filas que no sea null su valor recaudado y valor previsto podemos calcular su porcentaje.

df[filtro5 & ~filtro3 & filtro4 & df["valor previsto"]!=0].head(2)

def calcular_porcentaje(fila):
    if pd.isnull(fila['porcentaje recaudado']) and pd.notnull(fila['valor previsto']) and pd.notnull(fila['valor recaudado']) and fila['valor previsto']!=0:
        return (fila['valor recaudado'] / fila['valor previsto']) * 100
    return fila['porcentaje recaudado']

df['porcentaje recaudado'] = df.apply(calcular_porcentaje, axis=1)

# Algo a descendido, aunque no mucho.
(df.isnull().sum()/df.shape[0]*100)["porcentaje recaudado"]

- Las filas que tengan los valores a null y el porcentaje a 0 no aportan nada de valor y podemos quitarlas. Lo mismo ocurre si de las 4 columnas hay 3 a null y la restante a 0, esas las quitamos. O las columnas que tienen 3 ceros y un nan, los 0 si que nos aportaría un cierto valor pero no es nada reaslita que 3 de las 4 columnas sean 0, ahí claramente está habiendo un problema a la hora de recoger los datos que sería de vital importancial solucionarlo.

df.shape[0]

# filtro: nan nan nan 0
f1 = filtro1 & filtro2 & filtro3 & filtro8
# filtro: nan nan 0 nan
f2 = filtro1 & filtro2 & filtro7 & filtro4
# filtro: nan 0 nan nan
f3 = filtro1 & filtro6 & filtro3 & filtro4
# filtro: 0 nan nan nan
f4 = filtro5 & filtro2 & filtro3 & filtro4

# filtro: nan 0 0 0
f5 = filtro1 & filtro6 & filtro7 & filtro8
# filtro: 0 nan 0 0
f6 = filtro5 & filtro2 & filtro7 & filtro8
# filtro: 0 0 nan 0
f7 = filtro5 & filtro6 & filtro3 & filtro8
# filtro: 0 0 0 nan
f8 = filtro5 & filtro6 & filtro7 & filtro4
# filtro: 0 0 0 0 (podría dejarlos pero es que estos valores no sería realistas y harían ruido)
f9 = filtro5 & filtro6 & filtro7 & filtro8

filtros_combo = f1 | f2 | f3 | f4 | f5 | f6 | f7 | f8 | f9


df[f1|f2|f3|f4|f5|f6|f7|f8|f9].shape

df_final = df[~filtros_combo] #invierto el filtro para quedarme con las filas contrarias a las que lo cumplen.

df_final.shape

(f"Nos hemos quedado con un {np.round(df_final.shape[0]/df.shape[0]*100, 2)}% de la tabla original")

df_final.isnull().sum()/df_final.shape[0]*100

df_final.dtypes

### 7- Gestión de los nulos de fecha recaudación.


Un 2.3% de los datos de la columna son nulos, no hay forma de reemplazarlos por una fecha, pero por lo menos tenemos la columna de año y podriamos borara esta pero si quisieramos acceder a los meses ya no se podría por lo que por ahora vamos a dejarla como esta.

## Operaciones finales

df_final.reset_index(inplace=True)

df_final = df_final.drop(columns=['Unnamed: 0', 'index'])


df_final.head()

(df_final.isnull().sum()/df_final.shape[0]*100)

df_final.to_csv("../datos/tabla_limpia.csv")

# Conclusiones

1 - Hemos corregido los nombres de las organizaciones superiores a partir de los id y han quedado solo 10543 sin corregir lo que representa un 1.04% del total. Lo mismo hemos hecho para corregir los id a partir del nombre de la organización superioir, de nuevo queda un 1.04% sin corregir el cual representa las filas donde ambas columnas tienen valores nulos. porterioirmente hemos conseguido reducirlo a un 0.03%.

2 - Hemos corregido los nombres de las organizaciones a partir de los id y han quedado solo 905 sin corregir lo que representa un 0.1% del total. Lo mismo hemos hecho para corregir los id a partir del nombre de la organización, de nuevo queda un 0.1% sin corregir el cual representa las filas donde ambas columnas tienen valores nulos.

3 - Hemos corregido los nombres de las unidades gestoras a partir de los id y han quedado solo 614 sin corregir lo que representa un 0.06% del total. Lo mismo hemos hecho para corregir los id a partir del nombre de la unidad gestora, de nuevo queda un 0.06% sin corregir el cual representa las filas donde ambas columnas tienen valores nulos.

4 - Para las columnas 'categoria economica', 'origen ingreso', 'tipo ingreso', los nulos al no poder ser gestionados pues no sabemos que valor pueden tener los hemos asignado como "Sem informação" antes que dejarlos como nulos.

5 - Para las columnas de "valor previsto", "valor registrado", "valor recaudado", "porcentaje recaudado", aparecen una grandísima cantidad de valores nulos complicada de gestionar. Además también hay muchas filar con valores a 0 por lo que se ha tomado la decisión entratégica de filtrar las filas de tal forma que eliminariamos una cierta canidad de nulos que no aportaban ningún tipo de valor quedándono con el 96.2% de la tabla original.

6 - Los nulos de la fecha de recaudación no puedo gestionarlos de ninguna manera por lo que los mantengo.
