==> notebooks/3-eda.ipynb <==
# Fase 3: EDA

import pandas as pd
import numpy as np

import sys
import os

# Agregar el directorio 'src' al path
sys.path.append(os.path.abspath('../src'))

# Importar la función suma del módulo auxiliares
import soporte_limpieza as sl

df = pd.read_csv("../datos/tabla_limpia.csv", parse_dates=["fecha recaudacion"])
df.shape

### 1- Distribución de Ingresos por Categoría Económica:

df["categoria economica"].value_counts()

Analizar las categorías de ingresos más significativas y su participación en los ingresos totales.

Las más significativas son en cuanto a valor recaudado pues es lo que realmente nos interesa ya que es el dinerito que realmente entra a la caja.

df.groupby("categoria economica").sum(numeric_only=True)["valor recaudado"].sort_values(ascending=False)

l_aux = list(df.groupby("categoria economica").sum(numeric_only=True)["valor recaudado"].sort_values(ascending=False))
total_categorias =0
for valor in l_aux:
    total_categorias+=valor
total_categorias

np.round(df.groupby("categoria economica").sum(numeric_only=True)["valor recaudado"].sort_values(ascending=False)/24658378453467.75*100,2)

Hay dos categorías que son las que más recaudan claramente que son Receitas Correntes y de Capital con casi el 50% cada una de la recaudación.

Ahora queremos ver la diferencia promedio entre ingresos previstos y realizados por cada categoría.

df_comparacion = df.groupby("categoria economica").sum(numeric_only=True)[["valor previsto","valor recaudado"]]
df_comparacion["diferencia"] = df_comparacion['valor recaudado']-df_comparacion['valor previsto']

df_comparacion_rounded = df_comparacion / 1000000000000 # pasamos a billones
df_comparacion_rounded

df_comparacion["diferencia"].mean()

df_comparacion["diferencia"].median()

Por lo general, la diferencia entre recaudado y previsto suele ser negativa ya que hay dos valores negativos muy grandes, otro negativo mediano y dos positivos bajos. Por lo que de media se recauda medio billón menos de lo previsto.

### 2- Análisis Temporal:

df.columns

df_anios = df.groupby("anio recaudacion").sum(numeric_only=True)[["valor previsto", "valor registrado", "valor recaudado"]]
df_anios = df_anios / 1000000000000 # pasamos a billones

df_anios["recaudado-previsto"]= df_anios["valor recaudado"]-df_anios["valor previsto"]
df_anios["recaudado-registrado"]= df_anios["valor recaudado"]-df_anios["valor registrado"]

df_anios

df_anios["recaudado-previsto"].mean()

df_anios["recaudado-registrado"].mean()

En cuanto al valor recaudado podemos observar como ha ido aumentando a lo largo de los años, a excepción de 2017 donde hubo un pequeña caida respecto al año anterior. Además se pegó un gran salto entre 2019 y 2020 donde se recaudo medio billón más con respecto al año anterior. También se puede onservar como a partir de 2015 el valor que se refleja como registrado no es de apenas medio billon cuando se está recaudando en algunos casos hasta 5,6 o y veces más. Lo que nos plantea, ¿el dinero recaudado pero no resgitrado que pasa con él? ¿Se va a actualizar el registro o ese dinero desaparece? Porque es muchísimo dinero. Además, todo los años el previsto es mucho mayor que el recaudado, ¿por qué calculan siempre mal el previsto?

df_anios_meses = df.groupby([df["fecha recaudacion"].dt.year, df["fecha recaudacion"].dt.month]).sum(numeric_only=True)[["valor previsto", "valor registrado", "valor recaudado"]]

df_anios_meses_rounded = np.round(df_anios_meses/100000000000, 8)

pd.set_option("display.max_rows", None)
df_anios_meses_rounded

En cuanto a los 3 primero años que son 2013,2014 y 2015 solo hay registros del mes de diciembre. 

Para el resto de los años hay registro de todos los meses, se observa que el valor previsto por lo general solo tiene un valor coherente en el mes de enero, a excepción de 2021 que es en el mes de abril. Por lo general el resto de los meses suele estar a 0 excepción de uno donde hay un valor previsto bastante más bajo en comparación con el de enero. Esto lo que me parece indicar es que establecen un valor a principio de año de lo que se va a obtener en todo el año y puede que hagan algún pequeño ajuste a mediados de año.

En cuanto al valor registrado no hay datos que se puedan abstarear de forma genérica. cada mes por lo general se registra un valor a excepción de algunos que son 0, este valor puede ser positivo o negativo indicando que igual han surgido gatos y ese mes se han quedado en negativo. El valor registrado se podría estudiar más en profundidad pero ya de forma individual para cada año pero no parece seguir ningún patrón.

El valor recaudado para cada mes siempre es positivo y superior al valor registrado.

df_meses = df.groupby( df["fecha recaudacion"].dt.month).sum(numeric_only=True)[["valor previsto", "valor registrado", "valor recaudado"]]

df_meses_rounded = np.round(df_meses/1000000000, 8)

df_meses_rounded.sort_values(by = "valor recaudado", ascending=False)

Si agrupamos todos los valores de todos los años solo por meses se ve claramente como en diciembre es el mes que se recauda más con gran diferencia, esto se debe a que en 2013,2014 y 2015 solo se recauda en diciembre. el siguiente mes que más recauda es septiembre y octubre ya que puede que en estos mese es cuando ellos hagan su versión de declaración de la renta.

### 3-Identificación de Discrepancias por categorías.

Recordamos la tabla del primer paso:

df_comparacion = df.groupby("categoria economica").sum(numeric_only=True)[["valor previsto","valor recaudado"]]
df_comparacion["diferencia"] = df_comparacion['valor recaudado']-df_comparacion['valor previsto']

df_comparacion_rounded = df_comparacion / 1000000000000 # pasamos a billones
df_comparacion_rounded.sort_values(by="diferencia")

Las categorías con mayor diferencia entre lo recaudado y lo previsto son Receitas de Capital y Receitas Corrientes, estas son las dos mayores categorías de ingresos que obtienen casi el 50% cada una. Es complicado prever cuanto dinero van a obtener esta categorías recaudadoras sobre todo siendo las que gestionan la mitad de la recaudación del país. Estando presentes ante una subejecución por parte de ambas categorías. Sin embargo, Receitas de Capital - intra-orçamentárias y Receitas Correntes - intra-orçamentárias si que se aproximan mucho mejor al valor recaudado.

Tendría que estudiarse que esta ocurriendo con las dos primeras y por que se estiman recaudar más de un billón de lo que realmente se recauda.

Vamos a filtrar por años a ver si vemos algo más claro:

df_anio_categoria = df.groupby(["anio recaudacion", "categoria economica"]).sum(numeric_only=True)[["valor previsto","valor recaudado"]]
df_anio_categoria["diferencia"] = df_anio_categoria['valor recaudado']-df_anio_categoria['valor previsto']
df_anio_categoria = df_anio_categoria / 1000000000000 # pasamos a billones

# Solo quiero las dos preocupantes
categorias_a_seleccionar = ['Receitas Correntes', 'Receitas de Capital']
df_anio_categoria_filtrado = df_anio_categoria[df_anio_categoria.index.get_level_values('categoria economica').isin(categorias_a_seleccionar)]

df_anio_categoria_filtrado

No hay nada del todo claro, igual podemos decir que en 2019 y 2021 las receitas Correntes empezaron a acercarse un poco mejor al valor recaudado. Para receitas Capital en 2020 se estimo un poco de menos y llego 2021 y estimaron medio billón de más.

## Conclusiones

- En cuanto a categorías tenemos dos claras categorias que recaudan cada una casi el 50% de valor recaudado que son `Receitas Correntes` y `Receitas de Capital`. Estudiando la diferencia promedio entre los ingresos previstos y recaudados observamos que para estas do categorías siempre se prevee más de lo que se recauda y de hecho la media de las diferenias es negativa, es decir, de media se ha hecho una previsión erronea por categoría de entorno a medio billón.

- En cuanto al análisis temporal se observa que la recaudación ha ido en aumento cada año, habiendo un aumento significativo en 2020. Sin embargo, ocurre de nuevo que para todos los años el valor previsto es siempre mayor que el recaudado, recaudando de media un tercio me billón menos de lo esperado cada año. Por otro lado la diferencia entre recaudado y previsto es siempre positiva, y creciendo cada año, un claro indicativo de que no se están registrando todas las cuentas y media obteniendo cada año 2 billones mas de lo registrado, haciendonos la pregunta de, si este dinero no está registrado, dónde está. 

    En cuanto a los meses, 2013, 2014 y 2015 solo había valores del mes de diciembre por lo que no sorprendeque el mes que más aparece que se recauda si agrupamos por meses es diciembre, seguido de septiembre y octubre. El valor previsto parece ser establecido siempre en enero estando casi todos los demás meses a 0 con algún pequeño ajuste otro mes del año. En cuanto al valor registrado siempre es menor cada mes que el valor recaudado, siendo algunos meses incluso negativo.

- Hay dos clara categorías que presentan una subejecucción y son `Receitas Correntes` y `Receitas de Capital`, recaudando a lo largo de los años un billón menos de lo calculado. Por lo que un problema que se debe revisar como se está realizando esta estimación y por qué si todos los años fallán no aprenden de sus errores y mejoran algo.
